Added a post trigger ARM dwell (time.sleep()) to account for the amount of buffer time that would be needed for the various trigger modes.
For instance, if the trigger is set to MID, then a little more than half of the acquisition time is allotted.
For mode POST, a little more than the entire acq time is allotted.
For PRE, no additional dwell time for buffer filling is allotted as all data will acquired after the trigger is armed.

This should obviate the need for scripts to add explicit delays and guesses about why the DAQ isn't triggering in certain circumstances.