Added operation_complete() (*OPC?) to each SCPI operation.
There seems to be a race condition of this instrument with other instruments such as oscilloscopes.
The problem is very likely that the USB-GPIB dongle has a large buffer inside and closes out the USB transaction very early but the data into the instrument is likely being slow-walked at GPIB speed.
The *OPC? command, while often ignored by some instruments likes scopes that have separate interface and analog systems, will at least wrap this dongle buffer problem if that's the issue.

On my HP8110A, I was able to remove the goofy write delay on the trigger channel and my test ran without turd notes from the scope trigger and even ran a fair bit faster.