Evaluate other conditions in a database, and submit a FAIL entry if not all conditions are met.
expected_conditions is a string for the SELECT portion of a sqlite query that consists of boolean equations (e.g. vout1_meas > 0.5, enable_pin == 1, etc.)
report_conditions is a list of column names that will be included in the SELECT portion and will be included inthe conditions of the FAIL entry.
 where_clause is a string for the WHERE portion of a sqlite query that limits the rows that will be considered for the evaluation.